requirements.txt
python-dotenv>=1.0.0
google-auth>=2.23.0
google-api-python-client>=2.100.0
mysql-connector-python>=8.2.0
geoip2>=4.7.0
maxminddb>=2.2.0

